Use passenger-customizable
This gives us easy installations for Ruby, Node and ability to set up
background processes in the "correct" way.

http://phusion.github.io/baseimage-docker/ (which
passenger-customizable) is built on explains this better.

We might consider switching entirely to baseimage-docker for a smaller
image with less stuff running - ie passenger-customizable currently
exposes ports 80 and 443, albeit without any listening processes), which
we don't need.

However, for now this will do as it allows us to rely no their Ruby and
Node installation scripts.