Reduce number of constraints
Given that we have an alignment constrain for all nodes in a group we
only need a inequality constraint on a single node in each group in
order to move the entire group.

In other words; as long as the first nodes in each group are
horizontally separated all the other nodes will follow as they follow
the first node in their own group.